Move livrable editing form under editing item
